requirements.txt
langchain~=0.0.348
python-dotenv~=1.0.0
streamlit~=1.29.0
openai~=1.3.8
keyboard~=0.13.5
psutil~=5.9.4
office365==0.3.15
Office365_REST_Python_Client==2.5.3